pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "uma-geometry-optimizer"
version = "0.2.5"
description = "Simple UMA geometry optimization package with Fairchem UMA and Torch-Sim"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [
  { name = "Niklas Hölter", email = "niklas.hoelter@uni-muenster.de" },
]
keywords = ["chemistry", "geometry-optimization", "uma", "fairchem", "ase", "mlip"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "License :: OSI Approved :: MIT License",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Chemistry",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
  "numpy>=2.0,<2.3",
  "ase>=3.24",
  "torch>=2",
  "torch_sim_atomistic>=0.3.0",
  "morfeus-ml>=0.7",
  "rdkit>=2022.9.5",
  "fairchem-core>=2.7.1",
  "tables>=3.10.2",
  "hf_xet"
]

[project.optional-dependencies]
yaml = ["pyyaml>=6.0"]

[project.scripts]
uma-geom-opt = "uma_geometry_optimizer.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["uma_geometry_optimizer*"]
